==> notebooks/NB03_Build_A_Database_01.ipynb <==
These are the libraries you need to install:

<div style="width:60%;font-size:0.85em">

```bash
pip install pandas numpy lets-plot numerize tqdm sqlalchemy jupysql
```

import os
import sqlite3


import numpy as np
import pandas as pd

DATA_FOLDER = os.path.join('../data/')

CLEAN_FOLDER = os.path.join(DATA_FOLDER, 'raw')

#Puede faltar

conn = sqlite3.connect(os.path.join(DATA_FOLDER, 'football.db'))

### Table 1: Teams

df_teams = pd.read_csv('../data/raw/team_data.csv')
df_teams_xg = pd.read_csv('../data/raw/team_xg_data.csv')



df_teams_xg['League'] = df_teams_xg['Leagues']
df_teams_xg = df_teams_xg.drop('Leagues', axis='columns')


df_teams = pd.merge(df_teams, df_teams_xg, on=['Teams', 'League'], how='inner')
df_teams

df_teams.info()

def extract_team_id(url):
    parts = url.split('/')
    return parts[4]

df_teams['Team ID'] = df_teams['Links'].apply(extract_team_id)
df_teams

df_teams.to_sql('teams', conn, if_exists='replace', index=False)

pd.read_sql('SELECT * FROM teams LIMIT 10', conn)

### Table 2: Players

df_players = pd.read_csv('../data/raw/players_data.csv')
df_player_xg = pd.read_csv('../data/raw/players_xg_data.csv')

df_players

df_player_xg
df_player_xg.rename(columns = {'Aerial Duels Won':'Expected goals per shot'}, inplace = True)

df_players = pd.merge(df_players, df_player_xg, on=['Player Name'], how='inner') #LUEGO CAMBIAR
df_players.replace('-', 0, inplace=True)
df_players.rename(columns = {'Aerial Duels Won_y':'Expected goals per shot'}, inplace = True)
df_players

df_players['Team ID'] = df_players['Team Link'].apply(extract_team_id)
df_players

df_players.to_sql('players', conn, if_exists='replace', index=False)

Adding the corresponding team name to each player by the team id from the link

df_temporary = pd.read_sql(
'''
SELECT 
    players.*, 
    teams.[Teams]
FROM 
    players
INNER JOIN 
    teams
ON 
    players.[Team ID] = teams.[Team ID]
''', conn)

df_temporary.to_sql('players', conn, if_exists='replace', index=False)

pd.read_sql('SELECT * FROM players LIMIT 5', conn)

conn.close()
