==> notebooks/NB04_Data_Visualization_1.ipynb <==
## Graph 1

Put this into the terminal
```bash
pip install pandas matplotlib lets-plot ipykernel jupyterlab
```

import os
import pandas as pd
import sqlite3
import numpy as py

DATA_FOLDER = os.path.join('../data/')

RAW_FOLDER = os.path.join(DATA_FOLDER, 'raw')

conn = sqlite3.connect(os.path.join(DATA_FOLDER, 'football.db'))

pd.read_sql('SELECT  players.* LI', conn)

pd.read_sql('SELECT * FROM players LIMIT 5', conn)

df_players = pd.read_sql('SELECT * FROM players', conn)
df_teams = pd.read_sql('SELECT * FROM teams', conn)

df_teams

conn.close()
